pulled out action adapter into own file
